requirements.txt
# FastAPI e server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0

# Database
sqlalchemy==2.0.23

# Autenticazione
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0  # Compatibilità con passlib 1.7.4

# Modbus
pymodbus==3.6.9
pyserial>=3.5

# Templates
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0

